Moved utility to shared folder, updated all imports
